crawler - index comparison - factor in groups
